Move checkRule into its own file.
