+++ server/routes/createComment.ts
@@ -25,7 +25,7 @@ import {
   MixpanelCommunityInteractionPayload,
 } from '../../shared/analytics/types';
 import { SENDGRID_API_KEY } from '../config';
-import { checkRule } from '../util/ruleParser';
+import checkRule from '../util/checkRule';
 import RuleCache from '../util/ruleCache';
 
 const sgMail = require('@sendgrid/mail');
