+++ server/util/ruleParser.ts
@@ -13,7 +13,7 @@ export type RuleMetadata = {
   arguments: Array<{ name: string, description: string, type: RuleArgumentType }>;
 };
 
-type DefaultSchemaT = Record<string, Array<unknown>>;
+export type DefaultSchemaT = Record<string, Array<unknown>>;
 
 // the type that must be implemented by a RuleType
 export type IRuleType<SchemaT extends DefaultSchemaT> = {
@@ -99,32 +99,4 @@ export function validateRule(ruleSchema: any): DefaultSchemaT {
   const [ ruleId ] = Object.keys(ruleSchema);
   const ruleDef = RuleTypes[ruleId];
   return ruleDef.validateRule(ruleSchema);
-}
-
-export async function checkRule(
-  ruleCache: RuleCache,
-  models: DB,
-  ruleFk: number,
-  address: string,
-  transaction?: Transaction,
-): Promise<boolean> {
-  // always pass non-configured rules
-  if (!ruleFk) return true;
-
-  // check cache and return early if valid
-  if (await ruleCache.check(ruleFk, address)) return true;
-
-  // fetch the rule from the database by id
-  const ruleInstance = await models.Rule.findOne({ where: { id: ruleFk } });
-  if (!ruleInstance) return true;
-  const ruleJson = ruleInstance.rule as Record<string, DefaultSchemaT>;
-
-  // parse and run the rule according to its type
-  const [[ruleId, ruleSchema]] = Object.entries(ruleJson);
-  const ruleDef: RuleType = RuleTypes[ruleId];
-  const isValid = await ruleDef.check(ruleSchema, address, ruleInstance.chain_id, models, transaction);
-
-  // add new successes to cache
-  if (isValid) await ruleCache.add(ruleFk, address);
-  return isValid;
 }
